Moved async_requests_provider module
